refactor: substituido o infra/ por database/
