+++ src/adapter/driven/database/models/categoriaModel.ts
