+++ src/adapter/driven/database/models/faturaModel.ts
