+++ src/adapter/driven/database/models/produtoImagensModel.ts
