+++ src/adapter/driven/database/repository/categoriaDatabaseRepository.ts
