+++ src/adapter/driven/database/repository/faturaDatabaseRepository.ts
