+++ src/adapter/driven/database/repository/metodoPagamentoDatabaseRepository.ts
