+++ src/adapter/driven/database/seeders/cria-categorias.ts
